feat(cli): Add --version flag and print version on startup
This commit introduces functionality to display the application's version number, making it easier for users to identify which version they are running.

The project version is now passed from CMake to the C++ source code via the `GPUBENCH_VERSION` preprocessor definition. This allows the version to be displayed in two ways:

- A `--version` command-line flag has been added.
- The version number is now printed to the console upon application startup.